DefaultNetworkingService was moved to a separate module to better hide the CURRENT_PROTOCOL_VERSION constant from the rest of the code; the constant was renamed to PREFERRED_PROTOCOL_VERSION
