+++ p2p/src/net/default_backend/mod.rs
@@ -14,44 +14,31 @@
 // limitations under the License.
 
 pub mod backend;
+mod default_networking_service;
 mod peer;
 pub mod transport;
 pub mod types;
 
-use std::{marker::PhantomData, sync::Arc};
+use std::marker::PhantomData;
 
 use async_trait::async_trait;
-use common::time_getter::TimeGetter;
-use p2p_types::{services::Services, socket_address::SocketAddress};
-use tokio::{
-    sync::{mpsc, oneshot},
-    task::JoinHandle,
-};
+use tokio::sync::mpsc;
 
 use logging::log;
-use utils::atomics::SeqCstAtomicBool;
+use p2p_types::{services::Services, socket_address::SocketAddress};
 
 use crate::{
     error::P2pError,
     message::{PeerManagerMessage, SyncMessage},
     net::{
         self,
-        default_backend::transport::{TransportListener, TransportSocket},
         types::{ConnectivityEvent, SyncingEvent},
         ConnectivityService, MessagingService, NetworkingService,
     },
-    protocol::{ProtocolVersion, SupportedProtocolVersion},
     types::peer_id::PeerId,
-    P2pConfig, P2pEventHandler,
 };
 
-// The current protocol version.
-// Note that we intentionally keep this constant private, because most of the code should
-// not depend on its value.
-const CURRENT_PROTOCOL_VERSION: SupportedProtocolVersion = SupportedProtocolVersion::V1;
-
-#[derive(Debug)]
-pub struct DefaultNetworkingService<T: TransportSocket>(PhantomData<T>);
+pub use default_networking_service::DefaultNetworkingService;
 
 #[derive(Debug)]
 pub struct ConnectivityHandle<S: NetworkingService> {
@@ -104,103 +91,6 @@ impl Clone for MessagingHandle {
 #[derive(Debug)]
 pub struct SyncingEventReceiver {
     syncing_event_rx: mpsc::UnboundedReceiver<SyncingEvent>,
-}
-
-impl<T: TransportSocket> DefaultNetworkingService<T> {
-    #[allow(clippy::too_many_arguments)]
-    pub async fn start_with_version(
-        transport: <Self as NetworkingService>::Transport,
-        bind_addresses: Vec<SocketAddress>,
-        chain_config: Arc<common::chain::ChainConfig>,
-        p2p_config: Arc<P2pConfig>,
-        time_getter: TimeGetter,
-        shutdown: Arc<SeqCstAtomicBool>,
-        shutdown_receiver: oneshot::Receiver<()>,
-        subscribers_receiver: mpsc::UnboundedReceiver<P2pEventHandler>,
-        protocol_version: ProtocolVersion,
-    ) -> crate::Result<(
-        <Self as NetworkingService>::ConnectivityHandle,
-        <Self as NetworkingService>::MessagingHandle,
-        <Self as NetworkingService>::SyncingEventReceiver,
-        JoinHandle<()>,
-    )> {
-        let (cmd_tx, cmd_rx) = mpsc::unbounded_channel();
-        let (conn_event_tx, conn_event_rx) = mpsc::unbounded_channel();
-        let (syncing_event_tx, syncing_event_rx) = mpsc::unbounded_channel();
-        let socket = transport.bind(bind_addresses).await?;
-        let local_addresses = socket.local_addresses().expect("to have bind address available");
-
-        let backend = backend::Backend::<T>::new(
-            transport,
-            socket,
-            chain_config,
-            Arc::clone(&p2p_config),
-            time_getter.clone(),
-            cmd_rx,
-            conn_event_tx,
-            syncing_event_tx,
-            Arc::clone(&shutdown),
-            shutdown_receiver,
-            subscribers_receiver,
-            protocol_version,
-        );
-        let backend_task = tokio::spawn(async move {
-            match backend.run().await {
-                Ok(never) => match never {},
-                Err(P2pError::ChannelClosed) if shutdown.load() => {
-                    log::info!("Backend is shut down");
-                }
-                Err(e) => {
-                    shutdown.store(true);
-                    log::error!("Failed to run backend: {e}");
-                }
-            }
-        });
-
-        Ok((
-            ConnectivityHandle::new(local_addresses, cmd_tx.clone(), conn_event_rx),
-            MessagingHandle::new(cmd_tx),
-            SyncingEventReceiver { syncing_event_rx },
-            backend_task,
-        ))
-    }
-}
-
-#[async_trait]
-impl<T: TransportSocket> NetworkingService for DefaultNetworkingService<T> {
-    type Transport = T;
-    type ConnectivityHandle = ConnectivityHandle<Self>;
-    type MessagingHandle = MessagingHandle;
-    type SyncingEventReceiver = SyncingEventReceiver;
-
-    async fn start(
-        transport: Self::Transport,
-        bind_addresses: Vec<SocketAddress>,
-        chain_config: Arc<common::chain::ChainConfig>,
-        p2p_config: Arc<P2pConfig>,
-        time_getter: TimeGetter,
-        shutdown: Arc<SeqCstAtomicBool>,
-        shutdown_receiver: oneshot::Receiver<()>,
-        subscribers_receiver: mpsc::UnboundedReceiver<P2pEventHandler>,
-    ) -> crate::Result<(
-        Self::ConnectivityHandle,
-        Self::MessagingHandle,
-        Self::SyncingEventReceiver,
-        JoinHandle<()>,
-    )> {
-        Self::start_with_version(
-            transport,
-            bind_addresses,
-            chain_config,
-            p2p_config,
-            time_getter,
-            shutdown,
-            shutdown_receiver,
-            subscribers_receiver,
-            CURRENT_PROTOCOL_VERSION.into(),
-        )
-        .await
-    }
 }
 
 #[async_trait]
